[Portal] Separate LLM content extraction from search data extraction
