+++ apps/portal/package.json
@@ -8,7 +8,8 @@
     "dev": "next dev",
     "prebuild": "pnpm run create-index",
     "build": "next build",
-    "postbuild": "pnpm run extract-search-data && pnpm next-sitemap",
+    "postbuild": "pnpm run extract-search-data && pnpm run extract-llm-content && pnpm next-sitemap",
+    "extract-llm-content": "pnpm tsx scripts/extractLLMData.ts",
     "start": "next start",
     "lint": "biome check ./src && knip && eslint ./src",
     "fix": "biome check ./src --fix && eslint ./src --fix",
