+++ apps/portal/scripts/extractSearchData.ts
@@ -3,11 +3,8 @@ import { extractContent } from "../src/app/api/search/extraction";
 
 async function main() {
   const rootDir = process.cwd();
-  const { searchData, llmContent, llmFullContent } =
-    await extractContent(rootDir);
+  const { searchData } = await extractContent(rootDir);
   writeFileSync("./searchIndex.json", JSON.stringify(searchData, null, 2));
-  writeFileSync("./public/llms.txt", llmContent);
-  writeFileSync("./public/llms-full.txt", llmFullContent);
 }
 
 main();
