+++ apps/portal/src/app/api/search/extraction/index.ts
@@ -1,6 +1,4 @@
 import { readFile } from "node:fs/promises";
-import he from "he";
-import { NodeHtmlMarkdown } from "node-html-markdown";
 import {
   CommentNode as X_CommentNode,
   HTMLElement as X_HTMLElement,
@@ -15,23 +13,7 @@ import { trimExtraSpace } from "./trimExtraSpace";
 
 type ExtractedContent = {
   searchData: PageData[];
-  llmContent: string;
-  llmFullContent: string;
 };
-
-const llmsContentHeader = `\
-# thirdweb
-
-> Frontend, Backend, and Onchain tools to build complete web3 apps — on every EVM chain.
-
-## Docs
-`;
-
-const llmsFullContentHeader = `\
-# thirdweb
-
-> Frontend, Backend, and Onchain tools to build complete web3 apps — on every EVM chain.
-`;
 
 export async function extractContent(
   rootDir: string,
@@ -40,8 +22,6 @@ export async function extractContent(
   const htmlFiles = getFilesRecursive(nextOutputDir, "html");
 
   const pages: PageData[] = [];
-  let llmContent = "";
-  let llmFullContent = "";
 
   const noMainFound: string[] = [];
   const noH1Found: string[] = [];
@@ -85,16 +65,6 @@ export async function extractContent(
       if (pageData) {
         pages.push(pageData);
       }
-
-      // Extract LLM content
-      const { links, full } = extractPageLLMContent(
-        mainEl,
-        pageTitle,
-        filePath,
-        nextOutputDir,
-      );
-      llmContent += links ? `${links}\n` : "";
-      llmFullContent += full ? `${full}\n` : "";
     }),
   );
 
@@ -118,8 +88,6 @@ export async function extractContent(
 
   return {
     searchData: pages,
-    llmContent: `${llmsContentHeader}\n${llmContent}`,
-    llmFullContent: `${llmsFullContentHeader}\n${llmFullContent}`,
   };
 }
 
@@ -137,122 +105,6 @@ function extractPageSearchData(
     href: filePath.replace(nextOutputDir, "").replace(".html", ""),
     title: pageTitle ? trimExtraSpace(pageTitle) : "",
     sections: getPageSectionsForSearchIndex(main),
-  };
-}
-
-function extractPageLLMContent(
-  main: X_HTMLElement,
-  pageTitle: string | undefined,
-  filePath: string,
-  nextOutputDir: string,
-): { links: string; full: string } {
-  if (
-    main.getAttribute("data-noindex") === "true" ||
-    main.getAttribute("data-no-llm") === "true"
-  ) {
-    return { links: "", full: "" };
-  }
-
-  const htmlToMarkdown = new NodeHtmlMarkdown({
-    keepDataImages: false,
-    ignore: ["button"],
-    maxConsecutiveNewlines: 2,
-  });
-
-  let linksContent = "";
-  let fullContent = "";
-
-  const pageUrl = filePath.replace(nextOutputDir, "").replace(".html", "");
-
-  // Get first non-empty paragraph for description
-  const paragraphs = main.querySelectorAll("p");
-  let description = "";
-  for (const p of paragraphs) {
-    // skip noindex or no-llm paragraphs
-    if (p.closest("[data-noindex]") || p.closest("[data-no-llm]")) {
-      continue;
-    }
-
-    description = trimExtraSpace(htmlToMarkdown.translate(p.toString()));
-    if (description) {
-      break;
-    }
-  }
-
-  linksContent += `* [${pageTitle}](${pageUrl}): ${description || `Reference for ${pageTitle}`}`;
-
-  // Remove noindex and no-llm elements
-  const contentElements = main.querySelectorAll("*");
-  for (const element of contentElements) {
-    if (
-      element.getAttribute("data-noindex") === "true" ||
-      element.getAttribute("data-no-llm") === "true"
-    ) {
-      element.remove();
-    }
-  }
-
-  // Shift all heading elements to 1 step down (h1 > h2, h2 > h3, etc.)
-  const headings = main.querySelectorAll("h1, h2, h3, h4, h5, h6");
-  for (const heading of headings) {
-    const headingLevel = Number.parseInt(heading.tagName.replace("H", ""));
-    const newLevel = Math.min(headingLevel + 1, 6);
-    heading.tagName = `H${newLevel}`;
-  }
-
-  // prefix all the relative links with the `https://portal.thirdweb.com`
-  const links = main.querySelectorAll("a");
-  for (const link of links) {
-    const href = link.getAttribute("href");
-    if (href?.startsWith("/")) {
-      link.setAttribute("href", `https://portal.thirdweb.com${href}`);
-    }
-  }
-
-  // prefix all relative image links with the `https://portal.thirdweb.com`
-  const images = main.querySelectorAll("img");
-  for (const image of images) {
-    const src = image.getAttribute("src");
-    if (src?.startsWith("/")) {
-      image.setAttribute("src", `https://portal.thirdweb.com${src}`);
-    }
-  }
-
-  // for code blocks inside pre tags -> make them direct descendants of the pre tag
-  // so they are parsed as blocks by node-html-markdown + add language class
-  const preTags = main.querySelectorAll("pre");
-  for (const preTag of preTags) {
-    const codeBlock = parse(preTag.innerHTML.toString(), {
-      comment: false,
-      blockTextElements: {
-        pre: true,
-      },
-    }).querySelector("code");
-
-    if (codeBlock) {
-      const code = codeBlock
-        .querySelectorAll("div > div > div > div")
-        .map((x) => x.textContent)
-        .join("\n")
-        .trim();
-
-      const lang = codeBlock.getAttribute("lang");
-      codeBlock.textContent = code;
-
-      const newCodePreBlock = parse(
-        `<pre><code class=${lang ? `language-${lang}` : ""}>${he.encode(code)}</code></pre>`,
-      );
-
-      preTag.replaceWith(newCodePreBlock);
-    }
-  }
-
-  // Convert the cleaned HTML to markdown
-  fullContent += `${htmlToMarkdown.translate(main.toString())}`;
-
-  return {
-    links: linksContent,
-    full: fullContent,
   };
 }
